third_party/blink/web_tests/external/wpt/web-animations/interfaces/KeyframeEffect/constructor-expected.txt

This is a testharness.js-based test.
Found 1 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] A KeyframeEffect constructed without any KeyframeEffectOptions object
  assert_equals: default iterationComposite expected (string) "replace" but got (undefined) undefined
Harness: the test ran to completion.
